Update reference code to match editorial change
For the weeks and days part, we already used a shortcut instead of
looping, so we keep it the same as before. For the years and months
calculation, we change to the new definition of ISODateSurpasses.

UPSTREAM_COMMIT=f6e78d07dea1daf9e12c8c987cc12d4c9a2b2583