Polyfill: Add skip era mechanism
This is needed for dealing with the coptic era that
proposal-intl-era-monthcode drops, but is still present in old versions
of ICU as ERA0. We give it a Symbol era code so that it can never be
matched by user code.

UPSTREAM_COMMIT=a42ad8559512fcbf844373b0409f07f76611ae8b