Polyfill: Make Gregorian calendar arithmetic match ISO8601
While working on exhaustive test scripts, I discovered a bug in the
calendar code, which was a failure to apply the algorithm change in
PR #2535 to the Gregorian calendar. Meanwhile I added test262 tests
covering it (wrapping-at-end-of-month-gregorian.js). This change fixes
that bug.

This is just a bug in the reference code, not in the spec text.

UPSTREAM_COMMIT=1f6d6df82ed711ad042ff853d85f2269f9da4ac5