Polyfill: Make more useful out-of-range error messages
The existing error messages for PlainDate and PlainDateTime being
outside of the representable range were not very helpful for debugging,
because they did not show the value that was out of range. This change
adds that information to the error messages.

UPSTREAM_COMMIT=3414e61cb74dd108d21c543db7ed8554be2b82b4