Editorial: Revert inadvertent observable change
In #2989 I added checks to prevent calling GetUTCEpochNanoseconds with
out-of-range arguments, which, per ECMA-262, is legal but ill-defined.
(See https://github.com/tc39/ecma262/issues/1087). These checks should
not have changed any observable semantics, just caused some inputs to
throw that would have thrown anyway.

These checks later became CheckISODateRange.

However, in one of the checks I made a mistake: the
GetUTCEpochNanoseconds call in GetNamedTimeZoneEpochNanoseconds can
never trigger the ill-defined behaviour because the time zone is always
UTC in the time-zone-unaware algorithm given there. So the extra check
was not necessary, and in fact was incorrect for time-zone-aware
implementations, causing an observable change in semantics as they would
throw where they previously did not.

Since this was an unintended and unwanted change to observable
semantics, which was not approved in committee, we must revert it.

Web compatibility status: Firefox ships the correct behaviour. Chrome
currently ships the broken behaviour, but that has not yet been exposed
to the web unflagged. Safari has not implemented this part of the
proposal yet.

UPSTREAM_COMMIT=6274e5de91b499cc3063cf8372d1fb8ce5d027d0