Polyfill: Only cache for one calendar at a time
Almost always when we populate an object with a copy of another object's
cache, it is for calculations in the same calendar. So we can say that
any particular object's cache is for only one calendar, and then we
don't need to include the calendar ID in the lookup key.

The only time when we copied the cache and it _wasn't_ for the same
calendar, was after doing a withCalendar(). In this case it was useless
to copy the cache anyway, because the entries from the old calendar
would never be looked up anymore. So we change PlainDate and
PlainDateTime's withCalendar methods to clone a fresh ISO Date Record
object instead of reusing it.

This provides an additional small speedup because it makes the keys
shorter.

Also fixes outdated comments about OneObjectCache; it associates caches
with ISO Date Record objects, not Temporal objects.

See: #3153

UPSTREAM_COMMIT=111381871a2a8388a5806148efac5c35b9b0197e