Polyfill: Change format of eraInfo table
Object.entries is showing up as very hot on profiles of completeEraYear.
It seems we are calling Object.entries on this table repeatedly, so may
as well just store it in the entries format in the first place.

See: #3153

UPSTREAM_COMMIT=84453d0e544886f69bcd4c82f90287f071ff2ec2