Polyfill: Avoid computing Hebrew month code if we already have it
Sometimes when we reach this function, the month code is already present
on the calendarDate parameter. It's a small speedup to just use it if
it's there. (Not a large speedup, because calculating it doesn't involve
a DateTimeFormat operation.)

See: #3153

UPSTREAM_COMMIT=c8fa9d43dea225a9eced7e62c0d94f28c5a639ae