Polyfill: Handle change of formatted Śaka era name
The check for the Indian calendar BCE bug included the literal string
"Saka", which has been changed to "Śaka" in the latest ICU data. Ignore
the formatted era name and just check the formatted date to see if the
host is vulnerable to the bug.

UPSTREAM_COMMIT=314b112888f679156142ac15135131959ec32ff2