Polyfill: Don't print era in PlainMonthDay formatting
See the change in https://github.com/tc39/proposal-temporal/pull/3175
for details. `era` should not be added to the default options for
PlainMonthDay, because [[era]] isn't copied in step 12.b of
GetDateTimeFormat when defaults is MONTH-DAY.

The spec is correct here, this brings the reference code into alignment
with it.

UPSTREAM_COMMIT=0512c498552b07d3898cf2fa41acb360f4a004c6